DAOS-17914 dlck: check pools' metadata (#16999)
DLCK is a standalone utility so the user will have daos_server stopped (or crashed) and would like to check the pool for errors. Possibly because of already observed malfunctions. Hence, DLCK works in an environment in which any operation can crash it without warning. But its implemented in such a way to make sure it won't crash no matter what.

The PR introduces the first DLCK command: check. It's purpose is to validate pool's metadata down to the container's tree. It reports the progress of the sub-checks into dedicated log files and the final results or critical errors back to the user by printing it on stdout.

Signed-off-by: Jan Michalski <jan-marian.michalski@hpe.com>