DAOS-17442 engine: Improve start issue handling (#16376) (#16588)
We infer from the available log messages that bio_xsctxt_alloc might get stuck after "failed to init bdevs: -1003" for some unknown reason related to SPDK. This patch skips two while statements in bio_xsctxt_free that might be where we got stuck.

When not in the CR mode, a pool ignores nonexistent children, leading to states that the current code cannot handle well (yet). This patch restricts this CR behavior to the CR mode. (Ideally, we shall allow the healthy children to start and automatically exclude the nonexistent children in the future.)

To make pool start failures easier to spot, this patch adds a RAS event that is raised when we fail to start a pool.

Also, this patch

fixes "-1" xstream IDs in messages logged during engine setup,
fixes a few places where we should write "pool" instead of "pool service", including the name of an internal environment variable, and
adds some infrequent log messages to help future debugging of engine start issues.

Move the unusually late initialization of dss_tls_key and
daos_srv_modkey earlier, so that the TLS checks in server_id_cb will be
safe. Both objects are static, and there's no reason to initialize them
so late.

Signed-off-by: Li Wei <liwei@hpe.com>