DAOS-17358 rebuild: rebuilding test in obj_inflight_io_check (#16970)
This change removes the check of incoming client IO with
the ORF_REBUILDING_IO flag set.

Before the change, the intent of the check was to temporarily disallow
such IO while a rebuild was starting, when the PS leader engine first
distributes a fence / epoch value to all engines. The check caused
the IO to get an error -DER_UPDATE_AGAIN, causing the client to retry.

With forthcoming features like interactive/explicit rebuild control,
the rebuild *stop* case is negatively affected by this test, causing
all subsequent client IO after rebuild stops to block indefinitely,
until the rebuild is restarted by the administrator.

Other intermittent test failures, unrelated to the new feature, are also
being seen. Discussion with other developers have suggested that this
test in the engine code is not required.

Signed-off-by: Kenneth Cain <kenneth.cain@hpe.com>