DAOS-18162 mgmt: Fix pool create error handling (#17043)
It has been observed that, during a dmg pool create command, the
ds_mgmt_pool_query call in pool_create_fill_resp timed out, and
ds_mgmt_drpc_pool_create returned the error to the MS without cleaning
up the newly-created pool service replicas. The dmg pool create command
retried with the same pool UUID, but this time a different set of PS
replicas were chosen and created "on top of" the PS replicas created in
the first attempt. As a result, some of the PS replicas had been
bootstrapped with the earlier set of replicas, whereas some with the
later set of replicas---an inconsistent Raft cluster right from the
beginning. Later, such inconsistency was ignored (for unknown reasons)
by rdb_raft_update_node, and led to assertion failures in raft.

This patch works around the problem like this:

  - Fix ds_mgmt_drpc_pool_create to clean up the pool if
    pool_create_fill_resp returns an error. Shorten the timeout of the
    query call, for the PS is just created by the MS and shouldn't take
    five minutes respond to the query.

  - Tighten the check in ds_rsvc_start to refuse to create and bootstrap
    "on top of" an existing replica, just to be safe.

  - Fix rdb_raft_update_node to report unexpected replica states and
    abort, rather than silently ignoring it. This should prevent the
    assertion failure from being reached.

Signed-off-by: Li Wei <liwei@hpe.com>