DAOS-17629 cq: sys_logscan for backward old logs (#16849)
The new log format was proposed and implemented in #16772.
The cart_logparse.py script was modified in #16772 to support
the new log format only.
The daos_sys_logscan.py tool is based on cart_logparse.py script and
after #16772 landing it will no longer support the old log format.

Backporting the tool to release/2.6 enables analysis of the
old logs.

Signed-off-by: Tomasz Gromadzki <tomasz.gromadzki@hpe.com>