SRE-3104 ci: Use DAOS specific HTTPS_PROXY variable (#16496) (#16647)
The HTTPS_PROXY variable set in Jenkins may force any shell script to use a proxy server.
This is not expected, as we want to avoid using a proxy unless it is explicitly needed by
the DAOS tests or the build process.

The new variable DAOS_HTTPS_PROXY has been introduced and is used explicitly when internet access is required.

Signed-off-by: Tomasz Gromadzki <tomasz.gromadzki@hpe.com>