DAOS-17576 chk: keep orphan pool shard which status is DOWN or DOWNOUT - b26 (#16395)
When check engine verifies pool membership, it may discard the rank or
target which status is ‘DOWN' or 'DOWNOUT’ to release related space.
Such logic was fine before. But as incremental reintegration feature
is on the way, such logic needs to be improved; otherwise, if related
orphan pool shard is removed by check engine, subsequent reintegration
has to be started from the scratch instead of incremental work.


Signed-off-by: Fan Yong <fan.yong@hpe.com>