DAOS-17180 ddb: replace device offline (#16167) (#17223)
If a NVMe SSD namespace is failed to be discovered by SPDK, the engine
will fail to start due to missing device, so there is no way to do the
online device replacing.

To fix above missing SSD namespace issue, offline device replacing
could be used in following steps:

- Run 'ddb dev_list' to show all devices;
- Run 'ddb dev_replace' to replace the broken SSD with an unused new SSD
- Start engine;
- Run 'dmg storage replace' with same old and new device UUID to revive
  the replaced SSD (blobs will be auto created and reintegration will be
  triggered);

Limitation:
Offline device replacing can't handle the missing sys SSD in md-on-ssd
mode (RDB is located on sys SSD in md-on-ssd mode) so far.

Signed-off-by: Niu Yawei <yawei.niu@hpe.com>