DAOS-3985 control: Add ControlInterface to server config
By default, the control plane server binds to 0.0.0.0, which
means that it is listening to all addresses on all interfaces.
In some cases, the admin may prefer to specify a single interface
to be used for control plane traffic.

When control_iface is set in daos_server.yml, the server will
use the first IPv4 address on that interface as both the listen
address and the address recorded in the management database. If
the prometheus listener is configured, it will also use the first
address found for the control interface.

Features: control
Signed-off-by: Michael MacDonald <github@macdonald.cx>