DAOS-10139 rdb: Avoid reusing raft node IDs (#17178)
Overview

Module rdb uses ranks as Raft node IDs. If we destroy an RDB replica X
on a rank, and later create a replica Y of the same RDB on the same
rank, then X and Y share the same Raft node ID. That is potentially
unsafe, as explained in the Jira ticket.

The current patch extends the Raft node ID type with a 32-bit generation
number, so that X and Y above would have different Raft node IDs. In
module raft, the node ID type is

  raft_node_id_t: int --> long.

In module rdb, the replica ID type is

  d_rank_t --> rdb_replica_id_t (rank << 32 | generation).

The RDB layout and protocol versions need to change, as a result. The
current patch handles old pools (i.e., those created by DAOS 2.6) in the
old RDB layout. Upgrading from the old RDB layout to the new one is not
possible yet. We define the generation of an old-layout RDB replicas as
zero.

The Generation Allocator

We need to produce replica generation numbers when creating new
replicas. There are only two such cases:

  - [ds_pool_svc_dist_create] When creating an initial set of N
    replicas, we assign 1, 2, ..., N as generation numbers to the N
    replicas.

  - [ds_rsvc_add_replicas] When creating a new replica to be added to an
    existing RDB, we obtain a new generation number from an allocator,
    initialized to N + 1 in the previous case, stored in the RDB LC
    (i.e., replicated via Raft along with user data). See rdb_raft_init
    and rdb_alloc_replica_gen.

Hence, any two replicas have different generation numbers. The
implementation consists of the following pieces:

  - [rdb_lc_replica_gen_next] stores next generation persistently in the
    LC.
  - [rdb_alloc_replica_gen] is the RDB API function.
  - [rdb.d_gen_lock] does concurrency control for the allocation TX.
  - [rdb_path_attrs] refers to the parent of the root KVS, RDB_LC_ATTRS.

Because no Raft entry type allows us to access rdb_lc_replica_gen_next,
we essentially extends RDB TXs through rdb_path_attrs so that they are
able to access RDB_LC_ATTRS when used by module rdb itself.

The Protocol Changes

In module rdb, besides the implicit changes caused by raft_node_id_t,
each request and reply now passes the source and destination replica
IDs, so that we can check if a replica is receiving a request or reply
sent to a different generation on its rank.

In module ds_rsvc, we need to pass replica IDs in addition to ranks when
creating replicas. And, we need to pass RDB layout versions so that when
working with old pools we create new replicas with the old RDB layout.

Misc Changes

In ds_rsvc_start, when creating a new replica, we must ensure that any
existing replica has a matching replica ID---its generation may be
different. This area is still sloppy compared to the rest of the code.

In ds_rsvc_dist_start, ds_rsvc_start, and rdb_create, after adding the
layout version, there are too many create parameters, so we employ two
new create_params types to maintain readability.

Signed-off-by: Li Wei <liwei@hpe.com>