DAOS-18293 test: container/query_properties.yaml - Remove (1) from layout_type
The expected layout_type has been updated from POSIX (1)
to POSIX by PR17040. The PR removed the layout number (1)
because the value returned by get-prop needs to be able
to feed into set-prop, but it wouldn't work if the number
is there. Update the test to follow this change.

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: test_query_properties
Signed-off-by: Makito Kano <makito.kano@hpe.com>