DAOS-18387 test: recovery/ddb.py test_recovery_ddb_ls MD-on-SSD Support
To support MD-on-SSD for ddb, we need to support two commands.
ddb prov_mem and ddb ls with --db_path.

Update ddb_utils.py to support the new commands.

Add check_ram_used in recovery_utils.py to detect whether
the system is MD-on-SSD.

Update test_recovery_ddb_ls to support MD-on-SSD with the
new ddb commands.

We need to update the test yaml to run on MD-on-SSD/HW Medium,
but that will break other tests in ddb.py because they don't
support MD-on-SSD yet. Keep the original tests as ddb_pmem.py
and ddb_pmem.yaml and keep running them on VM (except
test_recovery_ddb_ls because that's updated in this PR).

Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium: false
Test-tag: test_recovery_ddb_ls DdbPMEMTest
Signed-off-by: Makito Kano <makito.kano@hpe.com>